fixed issues with score being int

+++ src/Elsa.CustomActivities/Handlers/WeightedCheckboxExpressionHandler.cs
@@ -90,7 +90,7 @@ namespace Elsa.CustomActivities.Handlers
             string value = await property.EvaluateFromExpressions<string>(evaluator, context, _logger, CancellationToken.None);
             bool isSingle = EvaluateIsSingle(property, evaluator, context);
             bool isPrePopulated = await EvaluatePrePopulated(property, evaluator, context);
-            int score = await EvaluateScore(property, evaluator, context);
+            decimal score = await EvaluateScore(property, evaluator, context);
 
             return new WeightedCheckboxRecord(identifier, value, isSingle, score, isPrePopulated);
 
@@ -135,12 +135,12 @@ namespace Elsa.CustomActivities.Handlers
             return false;
         }
 
-        public async Task<int> EvaluateScore(ElsaProperty property, IExpressionEvaluator evaluator, ActivityExecutionContext context)
+        public async Task<decimal> EvaluateScore(ElsaProperty property, IExpressionEvaluator evaluator, ActivityExecutionContext context)
         {
             if (property.Expressions!.ContainsKey(ScoringSyntaxNames.Score))
             {
                 string expression = property.Expressions[ScoringSyntaxNames.Score] ?? "0";
-                int score = await property.EvaluateFromExpressionsExplicit<int>(evaluator,
+                decimal score = await property.EvaluateFromExpressionsExplicit<decimal>(evaluator,
                     context, _logger,
                     expression,
                     SyntaxNames.Literal,
